ocr_demo_project/requirements_enhanced.txt
# 增强版智能体依赖包
# 安装: pip install -r requirements_enhanced.txt

# ============================================
# 核心依赖（原有）
# ============================================
openai>=1.0.0
langchain-openai>=0.0.2
python-dotenv>=0.19.0
tiktoken>=0.5.0
numpy>=1.24.0
tqdm>=4.65.0

# ============================================
# PDF处理
# ============================================
PyMuPDF>=1.23.0           # PDF基础处理（已有）
pymupdf>=1.23.0

# 高级表格提取（新增）
pdfplumber>=0.10.0        # 推荐：表格提取效果好
camelot-py[cv]>=0.11.0    # 可选：更精确的表格提取
beautifulsoup4>=4.12.0    # HTML解析（表格处理需要）
lxml>=4.9.0               # XML/HTML解析器
# 注意：camelot需要额外系统依赖
# Ubuntu/Debian: sudo apt-get install python3-tk ghostscript
# macOS: brew install ghostscript tcl-tk

# ============================================
# OCR支持（新增 - 默认启用）
# ============================================
# PaddleOCR（中文识别效果好，无需系统依赖）
paddleocr>=2.7.0          # PaddleOCR OCR引擎
paddlepaddle>=2.5.0       # PaddlePaddle深度学习框架（CPU版本）
# 如果有GPU，使用：
# paddlepaddle-gpu>=2.5.0  # GPU版本（更快）

# 图像处理（必需）
opencv-python>=4.8.0      # OpenCV - 图像预处理、旋转纠正、水印去除
opencv-contrib-python>=4.8.0  # OpenCV扩展功能
Pillow>=10.0.0            # 图像处理库

# 备选方案：Tesseract（可选）
pytesseract>=0.3.10       # Tesseract OCR接口
pdf2image>=1.16.3         # PDF转图片（Tesseract需要）

# Tesseract需要额外系统依赖：
# Ubuntu/Debian: 
#   sudo apt-get install tesseract-ocr tesseract-ocr-chi-sim tesseract-ocr-eng
# macOS: 
#   brew install tesseract tesseract-lang
# Windows:
#   下载安装包: https://github.com/UB-Mannheim/tesseract/wiki
#
# 注意：使用PaddleOCR无需安装Tesseract系统依赖

# ============================================
# 文本处理
# ============================================
markdown>=3.4.0           # Markdown处理

# ============================================
# 其他工具
# ============================================
requests>=2.31.0          # HTTP请求
chardet>=5.0.0            # 字符编码检测

# ============================================
# 可选依赖
# ============================================
# langdetect>=1.0.9       # 语言检测（可选）
# difflib                 # Python内置，无需安装
# hashlib                 # Python内置，无需安装
# json                    # Python内置，无需安装
# re                      # Python内置，无需安装

# ============================================
# 开发和测试（可选）
# ============================================
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0

